Fixed the orientation, this is what I get for using incorrect indexing in a loop (1..24) instead of (0..23)
This involved changes across two files. The orientation was fixed by iterating through the index in a different fashion by subtracting numbers. I also changed the direction in which the circles are generated as they were being drawn in the wrong direction after the previous orientation change.

Triangle Controller initialises board in the correct orientation