Okay so we have concurrency and a blocking queue. This means that when you press a button, it adds an event to the queue which is processed by the dispatched thread which in turn will cause an event to occur. This deals with the issue of potential recursion issues. The triangleController should not be passed to the Circles because the TriangleController is incharge of causing circles to be drawn. Imagine the potential situation in which you drawN circles inside the constructor of HeadCircle. It will repeatedly call itself and crash.
To describe the current changes, there is a thread that is dispatched which will forever catch events in the queue and process them until it receives a quit event. The Event class itself is not complete, but it shall contain events which will describe what the TriangleController needs to do next.