Another huge update
The minimax algorithm seems to work, it atleast returns the correct score for the correct move. No longer using expectimax.

SequencesOfMoves stores multiple pairs

A Pair stores a start and and end position

UniversalBoard might be used for taking inputs across networking and so on

Node can generate the correct children

A DiceRollHolder can generate a sequences of moves regardless of whether they are legal

Scorable's getScore takes a side now